v1.1.2
fixed some major bug
