+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lavamusic",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "lavalink music bot base in erelajs",
   "main": "src/index.js",
   "scripts": {
