+++ src/index.js
@@ -1,6 +1,22 @@
-const MusicBot = require("./structures/MusicClient");
+const MusicBot = require("./structures/Client");
 const client = new MusicBot();
 
 client.connect()
+
+process.on('unhandledRejection', (reason, p) => {
+    console.log(reason, p);
+});
+
+process.on('uncaughtException', (err, origin) => {
+    console.log(err, origin);
+});
+
+process.on('uncaughtExceptionMonitor', (err, origin) => {
+    console.log(err, origin);
+});
+
+process.on('multipleResolves', (type, promise, reason) => {
+    console.log(type, promise, reason);
+});
 
 module.exports = client;
